requirements.txt
psycopg2-binary
voyageai
anthropic
